Separate screen components into individual files
