+++ lib/screens/screens.dart
@@ -1,0 +1,3 @@
+export 'dashboard_screen.dart';
+export 'profiles_screen.dart';
+export 'extensions_screen.dart';
